[mod] rename color palette file #199
